// src/interfaces/review.interface.ts
/* eslint-disable no-unused-vars */
import { Model, Schema } from 'mongoose'

interface IReview {
  review: string
  rating: number
  createdAt: Date
  tour: Schema.Types.ObjectId
  user: Schema.Types.ObjectId
}

interface IReviewModal extends Model<IReview> {
  calcAverageRatings(tourId: Schema.Types.ObjectId): Promise<void>
}

export { IReview, IReviewModal }